+++ NEWS.md
@@ -6,6 +6,9 @@
 * `unnest()` gains a `.drop` argument which controls what happens to
   other list columns. By default, they're kept if the output doesn't require
   row duplication; otherwise they're dropped.
+  
+* `unnest()` now has `mutate()` semantics for `...` - this allows you to
+  unnest transformed columns more easily. (Previously it used select semantics).
 
 * `spread()` once again creates columns of mixed type when `convert = TRUE` (#118, @jennybc)